update main structure
